This is the file for the first lab.
Here I am proving simple proofs using the Coq/Roq methods. I used the intros method to use the natural deduction introduction rules. The apply method was used to use the natural deduction elimination rule. The assumption method was used to make an assumption on a line in the proof.